cut-and-paste
