+++ cockatrice/CMakeLists.txt
@@ -149,6 +149,8 @@ set(cockatrice_SOURCES
     src/client/tabs/tab_room.cpp
     src/client/tabs/tab_server.cpp
     src/client/tabs/tab_supervisor.cpp
+    src/client/tabs/deck_view_container.cpp
+    src/client/tabs/deck_view_container.h
     src/game/zones/table_zone.cpp
     src/client/tapped_out_interface.cpp
     src/client/ui/theme_manager.cpp
